+++ packages/editor/src/components/provider/use-block-editor-settings.js
@@ -92,6 +92,8 @@ const BLOCK_EDITOR_SETTINGS = [
 	'isRTL',
 	'locale',
 	'maxWidth',
+	'noteEmojibaseUrl',
+	'noteEmojiLabelOverrides',
 	'noteReactionEmojis',
 	'postContentAttributes',
 	'postsPerPage',